requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
aiosqlite
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
pytest
pytest-asyncio
python-dotenv
httpx